refactor: renamed folder names
