+++ lib/src/recipe_controller.dart
@@ -5,8 +5,8 @@ import 'package:http/http.dart' as http;
 import 'models/recipe.dart';
 import 'models/recipe_parsing_job.dart';
 import 'models/recipe_parsing_result.dart';
-import 'recipe-scripts/bianca_zapatka.dart';
-import 'recipe-scripts/kptncook.dart';
+import 'recipe_scripts/bianca_zapatka.dart';
+import 'recipe_scripts/kptncook.dart';
 
 final Map<String, RecipeParsingResult Function(Document, RecipeParsingJob)>
     _recipeParseMethodMap = {
