+++ test/recipe_scripts_tests/bianca_zapatka_test.dart
@@ -2,8 +2,8 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:html/dom.dart';
 import 'package:ingredient_collector/src/models/ingredient.dart';
 import 'package:ingredient_collector/src/models/recipe_parsing_job.dart';
-import 'package:ingredient_collector/src/recipe-scripts/bianca_zapatka.dart';
 import 'package:ingredient_collector/src/recipe_controller.dart';
+import 'package:ingredient_collector/src/recipe_scripts/bianca_zapatka.dart';
 
 import 'script_test_helper.dart';
